ARROW-4009: [CI] Run Valgrind and C++ code coverage in different builds
